+++ app/src/main/java/asdev/amansoft/com/colorclick/WorldHighScore.java
@@ -18,7 +18,7 @@ public class WorldHighScore extends AppCompatActivity {
     private int BackPressedState = 0;
     private TextView scoreText;
     private Button continueBtn;
-    private int score;
+    private long score;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -41,8 +41,8 @@ public class WorldHighScore extends AppCompatActivity {
     @Override
     protected void onStart() {
         super.onStart();
-        score = getIntent().getIntExtra("SCORE", 0);
-        scoreText.setText(Integer.toString(score));
+        score = getIntent().getLongExtra("SCORE", 0);
+        scoreText.setText(Long.toString(score));
 
         new Handler().postDelayed(new Runnable() {
             @Override
